Add api test to validate manifest against schema
This additional test retrieves the manifest via GET request and
validates that it complies with the OpenAPI Spec. It will be run in CI
as part of the MediaWiki API integration test framework.

Also fixes the manifest URL (/wikibase-manifest/ vs /wikibase/manifest/)
and the 'namespace_name' property in the local_entities section
according to the latest version of the specification.

Bug: T263661
Change-Id: Ibd1fad290837d73f80dac6559b756b26dcdef9b8