Dependabot -> Gerrit, Github Action
This Github Action will run on pullrequests that
are created by the github dependabot user.
The single commit in the pull request will be modified
(re authored) and pushed to gerrit as is.
The Change-Id of the gerrit patch will be some hashed
version of the depenedabot branch name, so that subsequent
changes (rebases etc) will push to the same PS on Gerrit.

This requires a secret, which I have already added to
the wikimedia github org for this repository
(And this repository only currently) so that a user (Addbot
currently) can submit the patch to gerrit.

In the case that a PR is closed on Github, the bot will
prefix the commit message with "Closed"
(At some point we might want to make this abandon changes.

Add .github/workflows/dependabot-gerrit.yml
Add .github/dependabot.yml (for package.json)

This has been tested on a fork of WikibaseManifest
And some changes can already be seen:
 - I68560efe98181c4e85a4840a8b2def0f3c04c482
 - I89e03a592aa72f301625187b496276326a6bd6ff
 - Ifd3a86d3dc1362f3e32136d8d00895c54e50122e

Bug: T244001
Change-Id: I6e27f726c8cf73a47fb34e8624d7341e7b19b079