Make LocalSourceEntityNamespacesFactory operate on DatabaseEntitySource
EntitySource is now an interface, and the getEntityNamespaceIds() method
only exists on the DatabaseEntitySource implementation, but not on
ApiEntitySource. Fortunately, WikibaseRepo::getLocalEntitySource() is
already declared to return a DatabaseEntitySource, so we just need to
adjust the parameter type in the constructor and the test, and it should
all work out (without having to support other EntitySource types in
LocalSourceEntityNamespacesFactory).

Bug: T291372
Change-Id: I8e1fa1322db9875fb4616c82ef94e70b7acb4ef2